+++ src/main/java/com/utn/ElBuenSabor/entities/DetallePedido.java
@@ -5,6 +5,8 @@ import jakarta.persistence.*;
 import lombok.*;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "detalle_pedido")
 @NoArgsConstructor
@@ -19,11 +21,11 @@ public class DetallePedido extends Base {
 
     @NotNull
     @Column(name = "subtotal", precision = 10, scale = 2)
-    private Double subtotal;
+    private BigDecimal subtotal;
 
     @NotNull
     @Column(name = "subtotal_costo", precision = 10, scale = 2)
-    private Double subtotalCosto;
+    private BigDecimal subtotalCosto;
 
     @ManyToOne()
     @JoinColumn(name = "id_articulo_manufacturado")
